Jogo class removed from main.py and moved to game.py
